Move setCenterPosition
